Created folders for each type to help with sorting
