+++ Backend/Queries/PostItemQuery.cs
@@ -1,7 +1,7 @@
 ﻿using Backend.Data;
 using Backend.Models;
 using Backend.Service;
-using Backend.Types;
+using Backend.Types.PostItem;
 using Microsoft.EntityFrameworkCore;
 
 namespace Backend.Queries
