+++ Backend/Types/Website/CreateWebsiteInput.cs
@@ -1,4 +1,4 @@
-namespace Backend.Types;
+namespace Backend.Types.Website;
 
 public class CreateWebsiteInput
 {
